chore: refactor the Accounts page into smaller components
