+++ src/app/Accounts/Accounts.tsx
@@ -1,417 +1,8 @@
-import {
-  PageSection,
-  PageSectionVariants,
-  Panel,
-  ToolbarItem,
-  Toolbar,
-  ToolbarContent,
-  TextContent,
-  Text,
-  SearchInput,
-  Spinner,
-  MenuToggle,
-  MenuContent,
-  MenuList,
-  MenuItem,
-  Menu,
-  Popper,
-  ToolbarToggleGroup,
-  ToolbarGroup,
-  ToolbarFilter,
-  Badge,
-} from '@patternfly/react-core';
-import { Table, Thead, Tr, Th, Tbody, Td } from '@patternfly/react-table';
-import React, { useEffect, useState } from 'react';
-import { Link, useLocation } from 'react-router-dom';
-import { getAccounts } from '../services/api';
-import { FilterIcon } from '@patternfly/react-icons';
-
-interface IAccounts {
-  name: string;
-  provider: string;
-  clusterCount: number;
-}
-interface TableToolbarProps {
-  setSearchValue: (value: string) => void;
-  searchValue: string;
-  setStatusSelection: (value: string) => void;
-  statusSelection: string;
-  setProviderSelections: (value: string[]) => void;
-  providerSelections: string[];
-  onSearchChange: (value: string) => void;
-}
-
-const TableToolbar: React.FunctionComponent<TableToolbarProps> = ({
-  searchValue,
-  setSearchValue,
-  setStatusSelection,
-  setProviderSelections,
-  providerSelections,
-}) => {
-  const onSearchChange = (value: string) => {
-    setSearchValue(value);
-  };
-
-  // Set up name search input
-  const searchInput = (
-    <SearchInput
-      placeholder="Filter by account name"
-      value={searchValue}
-      onChange={(_event, value) => onSearchChange(value)}
-      onClear={() => onSearchChange('')}
-    />
-  );
-
-  // Set up name input
-  const [isStatusMenuOpen, setIsStatusMenuOpen] = React.useState<boolean>(false);
-  const statusToggleRef = React.useRef<HTMLButtonElement>(null);
-  const statusMenuRef = React.useRef<HTMLDivElement>(null);
-  const handleStatusMenuKeys = (event: KeyboardEvent) => {
-    if (isStatusMenuOpen && statusMenuRef.current?.contains(event.target as Node)) {
-      if (event.key === 'Escape' || event.key === 'Tab') {
-        setIsStatusMenuOpen(!isStatusMenuOpen);
-        statusToggleRef.current?.focus();
-      }
-    }
-  };
-
-  const handleStatusClickOutside = (event: MouseEvent) => {
-    if (isStatusMenuOpen && !statusMenuRef.current?.contains(event.target as Node)) {
-      setIsStatusMenuOpen(false);
-    }
-  };
-
-  React.useEffect(() => {
-    window.addEventListener('keydown', handleStatusMenuKeys);
-    window.addEventListener('click', handleStatusClickOutside);
-    return () => {
-      window.removeEventListener('keydown', handleStatusMenuKeys);
-      window.removeEventListener('click', handleStatusClickOutside);
-    };
-  }, [isStatusMenuOpen, statusMenuRef]);
-
-  // Set up provider input
-  const [isProviderMenuOpen, setIsProviderMenuOpen] = React.useState<boolean>(false);
-  const providerToggleRef = React.useRef<HTMLButtonElement>(null);
-  const providerMenuRef = React.useRef<HTMLDivElement>(null);
-  const providerContainerRef = React.useRef<HTMLDivElement>(null);
-
-  const handleProviderMenuKeys = (event: KeyboardEvent) => {
-    if (isProviderMenuOpen && providerMenuRef.current?.contains(event.target as Node)) {
-      if (event.key === 'Escape' || event.key === 'Tab') {
-        setIsProviderMenuOpen(!isProviderMenuOpen);
-        providerToggleRef.current?.focus();
-      }
-    }
-  };
-
-  const handleProviderClickOutside = (event: MouseEvent) => {
-    if (isProviderMenuOpen && !providerMenuRef.current?.contains(event.target as Node)) {
-      setIsProviderMenuOpen(false);
-    }
-  };
-
-  React.useEffect(() => {
-    window.addEventListener('keydown', handleProviderMenuKeys);
-    window.addEventListener('click', handleProviderClickOutside);
-    return () => {
-      window.removeEventListener('keydown', handleProviderMenuKeys);
-      window.removeEventListener('click', handleProviderClickOutside);
-    };
-  }, [isProviderMenuOpen, providerMenuRef]);
-
-  const onProviderMenuToggleClick = (ev: React.MouseEvent) => {
-    ev.stopPropagation();
-    setTimeout(() => {
-      if (providerMenuRef.current) {
-        const firstElement = providerMenuRef.current.querySelector('li > button:not(:disabled)');
-        firstElement && (firstElement as HTMLElement).focus();
-      }
-    }, 0);
-    setIsProviderMenuOpen(!isProviderMenuOpen);
-  };
-
-  function onProviderMenuSelect(event: React.MouseEvent | undefined, itemId: string | number | undefined) {
-    if (typeof itemId === 'undefined') {
-      return;
-    }
-
-    const itemStr = itemId.toString();
-
-    setProviderSelections(
-      providerSelections.includes(itemStr)
-        ? providerSelections.filter((selection) => selection !== itemStr)
-        : [itemStr, ...providerSelections]
-    );
-  }
-
-  const providerToggle = (
-    <MenuToggle
-      ref={providerToggleRef}
-      onClick={onProviderMenuToggleClick}
-      isExpanded={isProviderMenuOpen}
-      {...(providerSelections.length > 0 && {
-        badge: <Badge isRead>{providerSelections.length}</Badge>,
-      })}
-      style={
-        {
-          width: '200px',
-        } as React.CSSProperties
-      }
-    >
-      Filter by provider
-    </MenuToggle>
-  );
-
-  const providerMenu = (
-    <Menu
-      ref={providerMenuRef}
-      id="attribute-search-provider-menu"
-      onSelect={onProviderMenuSelect}
-      selected={providerSelections}
-    >
-      <MenuContent>
-        <MenuList>
-          <MenuItem hasCheckbox isSelected={providerSelections.includes('AWS')} itemId="AWS">
-            AWS
-          </MenuItem>
-          <MenuItem hasCheckbox isSelected={providerSelections.includes('GCP')} itemId="GCP">
-            Google Cloud
-          </MenuItem>
-          <MenuItem hasCheckbox isSelected={providerSelections.includes('Azure')} itemId="Azure">
-            Azure
-          </MenuItem>
-        </MenuList>
-      </MenuContent>
-    </Menu>
-  );
-
-  const providerSelect = (
-    <div ref={providerContainerRef}>
-      <Popper
-        trigger={providerToggle}
-        triggerRef={providerToggleRef}
-        popper={providerMenu}
-        popperRef={providerMenuRef}
-        appendTo={providerContainerRef.current || undefined}
-        isVisible={isProviderMenuOpen}
-      />
-    </div>
-  );
-
-  // Set up attribute selector
-  const [activeAttributeMenu, setActiveAttributeMenu] = React.useState<'Account' | 'Provider'>('Account');
-  const [isAttributeMenuOpen, setIsAttributeMenuOpen] = React.useState(false);
-  const attributeToggleRef = React.useRef<HTMLButtonElement>(null);
-  const attributeMenuRef = React.useRef<HTMLDivElement>(null);
-  const attributeContainerRef = React.useRef<HTMLDivElement>(null);
-
-  const handleAttribueMenuKeys = (event: KeyboardEvent) => {
-    if (!isAttributeMenuOpen) {
-      return;
-    }
-    if (
-      attributeMenuRef.current?.contains(event.target as Node) ||
-      attributeToggleRef.current?.contains(event.target as Node)
-    ) {
-      if (event.key === 'Escape' || event.key === 'Tab') {
-        setIsAttributeMenuOpen(!isAttributeMenuOpen);
-        attributeToggleRef.current?.focus();
-      }
-    }
-  };
-
-  const handleAttributeClickOutside = (event: MouseEvent) => {
-    if (isAttributeMenuOpen && !attributeMenuRef.current?.contains(event.target as Node)) {
-      setIsAttributeMenuOpen(false);
-    }
-  };
-
-  React.useEffect(() => {
-    window.addEventListener('keydown', handleAttribueMenuKeys);
-    window.addEventListener('click', handleAttributeClickOutside);
-    return () => {
-      window.removeEventListener('keydown', handleAttribueMenuKeys);
-      window.removeEventListener('click', handleAttributeClickOutside);
-    };
-  }, [isAttributeMenuOpen, attributeMenuRef]);
-
-  const onAttributeToggleClick = (ev: React.MouseEvent) => {
-    ev.stopPropagation();
-    setTimeout(() => {
-      if (attributeMenuRef.current) {
-        const firstElement = attributeMenuRef.current.querySelector('li > button:not(:disabled)');
-        firstElement && (firstElement as HTMLElement).focus();
-      }
-    }, 0);
-    setIsAttributeMenuOpen(!isAttributeMenuOpen);
-  };
-
-  const attributeToggle = (
-    <MenuToggle
-      ref={attributeToggleRef}
-      onClick={onAttributeToggleClick}
-      isExpanded={isAttributeMenuOpen}
-      icon={<FilterIcon />}
-    >
-      {activeAttributeMenu}
-    </MenuToggle>
-  );
-  const attributeMenu = (
-    // eslint-disable-next-line no-console
-    <Menu
-      ref={attributeMenuRef}
-      onSelect={(_ev, itemId) => {
-        setActiveAttributeMenu(itemId?.toString() as 'Account' | 'Provider');
-        setIsAttributeMenuOpen(!isAttributeMenuOpen);
-      }}
-    >
-      <MenuContent>
-        <MenuList>
-          <MenuItem itemId="Account">Account</MenuItem>
-          <MenuItem itemId="Provider">Provider</MenuItem>
-        </MenuList>
-      </MenuContent>
-    </Menu>
-  );
-
-  const attributeDropdown = (
-    <div ref={attributeContainerRef}>
-      <Popper
-        trigger={attributeToggle}
-        triggerRef={attributeToggleRef}
-        popper={attributeMenu}
-        popperRef={attributeMenuRef}
-        appendTo={attributeContainerRef.current || undefined}
-        isVisible={isAttributeMenuOpen}
-      />
-    </div>
-  );
-
-  return (
-    <Toolbar
-      id="attribute-search-filter-toolbar"
-      clearAllFilters={() => {
-        setSearchValue('');
-        setStatusSelection('');
-        setProviderSelections([]);
-      }}
-    >
-      <ToolbarContent>
-        <ToolbarToggleGroup toggleIcon={<FilterIcon />} breakpoint="xl">
-          <ToolbarGroup variant="filter-group">
-            <ToolbarItem>{attributeDropdown}</ToolbarItem>
-            <ToolbarFilter
-              chips={searchValue !== '' ? [searchValue] : ([] as string[])}
-              deleteChip={() => setSearchValue('')}
-              deleteChipGroup={() => setSearchValue('')}
-              categoryName="Name"
-              showToolbarItem={activeAttributeMenu === 'Account'}
-            >
-              {searchInput}
-            </ToolbarFilter>
-
-            <ToolbarFilter
-              chips={providerSelections}
-              deleteChip={(category, chip) => onProviderMenuSelect(undefined, chip as string)}
-              deleteChipGroup={() => setProviderSelections([])}
-              categoryName="Provider"
-              showToolbarItem={activeAttributeMenu === 'Provider'}
-            >
-              {providerSelect}
-            </ToolbarFilter>
-          </ToolbarGroup>
-        </ToolbarToggleGroup>
-      </ToolbarContent>
-    </Toolbar>
-  );
-};
-
-const AccountTable: React.FunctionComponent<{
-  searchValue: string;
-  statusFilter: string | null;
-  cloudProviderFilter: string | null;
-  providerSelections;
-}> = ({ searchValue, cloudProviderFilter, providerSelections }) => {
-  const [accountData, setAccountData] = useState<IAccounts[] | []>([]);
-  const [filteredData, setFilteredData] = useState<IAccounts[] | []>([]);
-  const [loading, setLoading] = useState(true);
-
-  useEffect(() => {
-    const fetchData = async () => {
-      setLoading(true);
-      try {
-        const fetchedAccounts = await getAccounts();
-        setAccountData(fetchedAccounts.accounts);
-      } catch (error) {
-        console.error('Error fetching data:', error);
-        console.error('Error fetching data:', error);
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchData();
-  }, []);
-
-  useEffect(() => {
-    let filtered = accountData.filter((account) => account.name.toLowerCase().includes(searchValue.toLowerCase()));
-
-    if (providerSelections.length > 0) {
-      filtered = filtered.filter((cluster) => providerSelections.includes(cluster.provider));
-    }
-
-    if (cloudProviderFilter) {
-      filtered = filtered.filter((cluster) => cluster.provider === cloudProviderFilter);
-    }
-
-    setFilteredData(filtered);
-  }, [searchValue, accountData, providerSelections, cloudProviderFilter]);
-
-  const columnNames = {
-    name: 'Name',
-    cloudProvider: 'Cloud Provider',
-    clusterCount: 'Cluster Count',
-  };
-
-  return (
-    <React.Fragment>
-      {loading ? (
-        <div
-          style={{
-            display: 'flex',
-            justifyContent: 'center',
-            alignItems: 'center',
-            height: '100vh',
-          }}
-        >
-          <Spinner size="xl" />
-        </div>
-      ) : (
-        <Table aria-label="Simple table">
-          <Thead>
-            <Tr>
-              <Th>{columnNames.name}</Th>
-              <Th>{columnNames.cloudProvider}</Th>
-              <Th>{columnNames.clusterCount}</Th>
-            </Tr>
-          </Thead>
-          <Tbody>
-            {filteredData.map((account) => (
-              <Tr key={account.name}>
-                <Td dataLabel={columnNames.name}>
-                  <Link to={`/accounts/${account.name}`}>{account.name}</Link>
-                </Td>
-                <Td dataLabel={columnNames.cloudProvider}>{account.provider}</Td>
-                <Td dataLabel={columnNames.clusterCount}>{account.clusterCount}</Td>
-              </Tr>
-            ))}
-          </Tbody>
-        </Table>
-      )}
-    </React.Fragment>
-  );
-};
+import { PageSection, PageSectionVariants, Panel, TextContent, Text } from '@patternfly/react-core';
+import React, { useState } from 'react';
+import { useLocation } from 'react-router-dom';
+import AccountsToolbar from './components/AccountsToolbar';
+import AccountsTable from './components/AccountsTable';
 
 const Accounts: React.FunctionComponent = () => {
   const [searchValue, setSearchValue] = useState<string>('');
@@ -419,12 +10,6 @@ const Accounts: React.FunctionComponent = () => {
   const location = useLocation();
   const queryParams = new URLSearchParams(location.search);
   const cloudProviderFilter = queryParams.get('cloudProvider');
-  console.log(queryParams.get('cloudProvider'));
-  console.log(cloudProviderFilter);
-
-  function setStatusSelection(value: string): void {
-    throw new Error('Function not implemented.');
-  }
 
   return (
     <React.Fragment>
@@ -435,16 +20,14 @@ const Accounts: React.FunctionComponent = () => {
       </PageSection>
       <PageSection variant={PageSectionVariants.light} isFilled>
         <Panel>
-          <TableToolbar
+          <AccountsToolbar
             onSearchChange={setSearchValue}
             setSearchValue={setSearchValue}
             searchValue={searchValue}
-            setStatusSelection={setStatusSelection}
-            statusSelection={''}
             setProviderSelections={setProviderSelections}
             providerSelections={providerSelections}
           />
-          <AccountTable
+          <AccountsTable
             searchValue={searchValue}
             cloudProviderFilter={cloudProviderFilter}
             providerSelections={providerSelections}
