+++ src/app/Accounts/types.ts
@@ -1,14 +1,20 @@
 export interface AccountsTableItem {
+  clusterCount: number;
   name: string;
   provider: string;
-  clusterCount: number;
 }
 
 export interface AccountsTableProps {
+  cloudProviderFilter: string | null;
+  providerSelections;
   searchValue: string;
+  statusFilter: string | null;
 }
 
 export interface AccountsToolbarProps {
-  value?: string;
+  searchValue: string;
+  setSearchValue: (value: string) => void;
+  providerSelections: string[];
+  setProviderSelections: (value: string[]) => void;
   onSearchChange: (value: string) => void;
 }
